+++ cherab/core/tests/test_lineshapes.py
@@ -1,6 +1,6 @@
-# Copyright 2016-2018 Euratom
-# Copyright 2016-2018 United Kingdom Atomic Energy Authority
-# Copyright 2016-2018 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
+# Copyright 2016-2023 Euratom
+# Copyright 2016-2023 United Kingdom Atomic Energy Authority
+# Copyright 2016-2023 Centro de Investigaciones Energéticas, Medioambientales y Tecnológicas
 #
 # Licensed under the EUPL, Version 1.1 or – as soon they will be approved by the
 # European Commission - subsequent versions of the EUPL (the "Licence");
@@ -26,11 +26,12 @@ from raysect.core import Point3D, Vector3D
 from raysect.core.math.function.float import Arg1D, Constant1D
 from raysect.optical import Spectrum
 
-from cherab.core import Line
+from cherab.core import Beam, Line, AtomicData
 from cherab.core.math.integrators import GaussianQuadrature
 from cherab.core.atomic import deuterium, nitrogen, ZeemanStructure
 from cherab.tools.plasmas.slab import build_constant_slab_plasma
 from cherab.core.model import GaussianLine, MultipletLineShape, StarkBroadenedLine, ZeemanTriplet, ParametrisedZeemanTriplet, ZeemanMultiplet
+from cherab.core.model import BeamEmissionMultiplet
 
 
 ATOMIC_MASS = 1.66053906660e-27
@@ -38,6 +39,9 @@ ELEMENTARY_CHARGE = 1.602176634e-19
 SPEED_OF_LIGHT = 299792458.0
 BOHR_MAGNETON = 5.78838180123e-5  # in eV/T
 HC_EV_NM = 1239.8419738620933  # (Planck constant in eV s) x (speed of light in nm/s)
+VACUUM_PERMITTIVITY = 8.8541878128e-12
+ELECTRON_REST_MASS = 9.1093837015e-31
+PLANCK_CONSTANT = 6.62607015e-34
 
 
 class TestLineShapes(unittest.TestCase):
@@ -46,13 +50,19 @@ class TestLineShapes(unittest.TestCase):
                       (nitrogen, 1, 1.e17, 10., Vector3D(1.e4, 5.e4, 0))]
     plasma = build_constant_slab_plasma(length=1, width=1, height=1, electron_density=1e19, electron_temperature=20.,
                                         plasma_species=plasma_species, b_field=Vector3D(0, 5., 0))
+    atomic_data = AtomicData()
+    beam = Beam()
+    beam.plasma = plasma
+    beam.energy = 60000
+    beam.temperature = 10
+    beam.element = deuterium
 
     def test_gaussian_line(self):
         # setting up a line shape model
         line = Line(deuterium, 0, (3, 2))  # D-alpha line
         target_species = self.plasma.composition.get(line.element, line.charge)
         wavelength = 656.104
-        gaussian_line = GaussianLine(line, wavelength, target_species, self.plasma)
+        gaussian_line = GaussianLine(line, wavelength, target_species, self.plasma, self.atomic_data)
 
         # spectrum parameters
         min_wavelength = wavelength - 0.5
@@ -87,7 +97,7 @@ class TestLineShapes(unittest.TestCase):
         target_species = self.plasma.composition.get(line.element, line.charge)
         multiplet = [[403.509, 404.132, 404.354, 404.479, 405.692], [0.205, 0.562, 0.175, 0.029, 0.029]]
         wavelength = 404.21
-        multiplet_line = MultipletLineShape(line, wavelength, target_species, self.plasma, multiplet)
+        multiplet_line = MultipletLineShape(line, wavelength, target_species, self.plasma, self.atomic_data, multiplet)
 
         # spectrum parameters
         min_wavelength = min(multiplet[0]) - 0.5
@@ -123,7 +133,7 @@ class TestLineShapes(unittest.TestCase):
         line = Line(deuterium, 0, (3, 2))  # D-alpha line
         target_species = self.plasma.composition.get(line.element, line.charge)
         wavelength = 656.104
-        triplet = ZeemanTriplet(line, wavelength, target_species, self.plasma)
+        triplet = ZeemanTriplet(line, wavelength, target_species, self.plasma, self.atomic_data)
 
         # spectrum parameters
         min_wavelength = wavelength - 0.5
@@ -174,7 +184,7 @@ class TestLineShapes(unittest.TestCase):
         line = Line(deuterium, 0, (3, 2))  # D-alpha line
         target_species = self.plasma.composition.get(line.element, line.charge)
         wavelength = 656.104
-        triplet = ParametrisedZeemanTriplet(line, wavelength, target_species, self.plasma)
+        triplet = ParametrisedZeemanTriplet(line, wavelength, target_species, self.plasma, self.atomic_data)
 
         # spectrum parameters
         min_wavelength = wavelength - 0.5
@@ -192,7 +202,7 @@ class TestLineShapes(unittest.TestCase):
             spectrum[pol] = triplet.add_line(radiance, point, direction, spectrum[pol])
 
         # validating
-        alpha, beta, gamma = triplet.LINE_PARAMETERS_DEFAULT[line]
+        alpha, beta, gamma = self.atomic_data.zeeman_triplet_parameters(line)
         temperature = target_species.distribution.effective_temperature(point.x, point.y, point.z)
         velocity = target_species.distribution.bulk_velocity(point.x, point.y, point.z)
         sigma = np.sqrt(temperature * ELEMENTARY_CHARGE / (line.element.atomic_weight * ATOMIC_MASS)) * wavelength / SPEED_OF_LIGHT
@@ -233,7 +243,7 @@ class TestLineShapes(unittest.TestCase):
         sigma_minus_components = [(HC_EV_NM / (photon_energy + BOHR_MAGNETON * Arg1D()), Constant1D(0.5))]
 
         zeeman_structure = ZeemanStructure(pi_components, sigma_plus_components, sigma_minus_components)
-        multiplet = ZeemanMultiplet(line, wavelength, target_species, self.plasma, zeeman_structure)
+        multiplet = ZeemanMultiplet(line, wavelength, target_species, self.plasma, self.atomic_data, zeeman_structure)
 
         # spectrum parameters
         min_wavelength = wavelength - 0.5
@@ -285,7 +295,7 @@ class TestLineShapes(unittest.TestCase):
         target_species = self.plasma.composition.get(line.element, line.charge)
         wavelength = 656.104
         integrator = GaussianQuadrature(relative_tolerance=1.e-5)
-        stark_line = StarkBroadenedLine(line, wavelength, target_species, self.plasma, integrator=integrator)
+        stark_line = StarkBroadenedLine(line, wavelength, target_species, self.plasma, self.atomic_data, integrator=integrator)
 
         # spectrum parameters
         min_wavelength = wavelength - 0.2
@@ -300,24 +310,162 @@ class TestLineShapes(unittest.TestCase):
         spectrum = stark_line.add_line(radiance, point, direction, spectrum)
 
         # validating
-        cij, aij, bij = stark_line.STARK_MODEL_COEFFICIENTS_DEFAULT[line]
+        velocity = target_species.distribution.bulk_velocity(point.x, point.y, point.z)
+        doppler_factor = (1 + velocity.dot(direction.normalise()) / SPEED_OF_LIGHT)
+
+        b_field = self.plasma.b_field(point.x, point.y, point.z)
+        b_magn = b_field.length
+        photon_energy = HC_EV_NM / wavelength
+        wl_sigma_plus = HC_EV_NM / (photon_energy - BOHR_MAGNETON * b_magn)
+        wl_sigma_minus = HC_EV_NM / (photon_energy + BOHR_MAGNETON * b_magn)
+        cos_sqr = (b_field.dot(direction.normalise()) / b_magn)**2
+        sin_sqr = 1. - cos_sqr
+
+        # Gaussian parameters
+        temperature = target_species.distribution.effective_temperature(point.x, point.y, point.z)
+        sigma = np.sqrt(temperature * ELEMENTARY_CHARGE / (line.element.atomic_weight * ATOMIC_MASS)) * wavelength / SPEED_OF_LIGHT
+        fwhm_gauss = 2 * np.sqrt(2 * np.log(2)) * sigma
+
+        # Lorentzian parameters
+        cij, aij, bij = self.atomic_data.stark_model_coefficients(line)
         ne = self.plasma.electron_distribution.density(point.x, point.y, point.z)
         te = self.plasma.electron_distribution.effective_temperature(point.x, point.y, point.z)
-        lambda_1_2 = cij * ne**aij / (te**bij)
+        fwhm_lorentz = cij * ne**aij / (te**bij)
 
-        lorenzian_cutoff_gamma = 50
-        stark_norm_coeff = 4 * lorenzian_cutoff_gamma * hyp2f1(0.4, 1, 1.4, -(2 * lorenzian_cutoff_gamma)**2.5)
-        norm = (0.5 * lambda_1_2)**1.5 / stark_norm_coeff
+        # Total FWHM
+        if fwhm_gauss <= fwhm_lorentz:
+            fwhm_poly_coeff = [1., 0, 0.57575, 0.37902, -0.42519, -0.31525, 0.31718]
+            fwhm_ratio = fwhm_gauss / fwhm_lorentz
+            fwhm_full = fwhm_lorentz * np.poly1d(fwhm_poly_coeff[::-1])(fwhm_ratio)
+        else:
+            fwhm_poly_coeff = [1., 0.15882, 1.04388, -1.38281, 0.46251, 0.82325, -0.58026]
+            fwhm_ratio = fwhm_lorentz / fwhm_gauss
+            fwhm_full = fwhm_gauss * np.poly1d(fwhm_poly_coeff[::-1])(fwhm_ratio)
 
         wavelengths, delta = np.linspace(min_wavelength, max_wavelength, bins + 1, retstep=True)
 
-        def stark_lineshape(x):
-            return norm / ((np.abs(x - wavelength))**2.5 + (0.5 * lambda_1_2)**2.5)
+        # Gaussian part
+        temp = 2 * np.sqrt(np.log(2)) / fwhm_full
+        erfs = erf((wavelengths - wavelength * doppler_factor) * temp)
+        gaussian = 0.25 * sin_sqr * (erfs[1:] - erfs[:-1]) / delta
+        erfs = erf((wavelengths - wl_sigma_plus * doppler_factor) * temp)
+        gaussian += 0.5 * (0.25 * sin_sqr + 0.5 * cos_sqr) * (erfs[1:] - erfs[:-1]) / delta
+        erfs = erf((wavelengths - wl_sigma_minus * doppler_factor) * temp)
+        gaussian += 0.5 * (0.25 * sin_sqr + 0.5 * cos_sqr) * (erfs[1:] - erfs[:-1]) / delta
+
+        # Lorentzian part
+        lorenzian_cutoff_gamma = 50
+        stark_norm_coeff = 4 * lorenzian_cutoff_gamma * hyp2f1(0.4, 1, 1.4, -(2 * lorenzian_cutoff_gamma)**2.5)
+        norm = (0.5 * fwhm_full)**1.5 / stark_norm_coeff
+
+        def stark_lineshape_pi(x):
+            return norm / ((np.abs(x - wavelength * doppler_factor))**2.5 + (0.5 * fwhm_full)**2.5)
+
+        def stark_lineshape_sigma_plus(x):
+            return norm / ((np.abs(x - wl_sigma_plus * doppler_factor))**2.5 + (0.5 * fwhm_full)**2.5)
+
+        def stark_lineshape_sigma_minus(x):
+            return norm / ((np.abs(x - wl_sigma_minus * doppler_factor))**2.5 + (0.5 * fwhm_full)**2.5)
+
+        weight_poly_coeff = [5.14820e-04, 1.38821e+00, -9.60424e-02, -3.83995e-02, -7.40042e-03, -5.47626e-04]
+        lorentz_weight = np.exp(np.poly1d(weight_poly_coeff[::-1])(np.log(fwhm_lorentz / fwhm_full)))
 
         for i in range(bins):
-            stark_bin = quadrature(stark_lineshape, wavelengths[i], wavelengths[i + 1], rtol=integrator.relative_tolerance)[0] / delta
-            self.assertAlmostEqual(stark_bin, spectrum.samples[i], delta=1e-9,
+            lorentz_bin = 0.5 * sin_sqr * quadrature(stark_lineshape_pi, wavelengths[i], wavelengths[i + 1],
+                                                     rtol=integrator.relative_tolerance)[0] / delta
+            lorentz_bin += (0.25 * sin_sqr + 0.5 * cos_sqr) * quadrature(stark_lineshape_sigma_plus, wavelengths[i], wavelengths[i + 1],
+                                                                         rtol=integrator.relative_tolerance)[0] / delta
+            lorentz_bin += (0.25 * sin_sqr + 0.5 * cos_sqr) * quadrature(stark_lineshape_sigma_minus, wavelengths[i], wavelengths[i + 1],
+                                                                         rtol=integrator.relative_tolerance)[0] / delta
+            ref_value = lorentz_bin * lorentz_weight + gaussian[i] * (1. - lorentz_weight)
+            self.assertAlmostEqual(ref_value, spectrum.samples[i], delta=1e-9,
                                    msg='StarkBroadenedLine.add_line() method gives a wrong value at {} nm.'.format(wavelengths[i]))
+
+    def test_beam_emission_multiplet(self):
+        # Test MSE line shape
+        # setting up a line shape model
+        line = Line(deuterium, 0, (3, 2))  # D-alpha line
+        wavelength = 656.104
+        mse_line = BeamEmissionMultiplet(line, wavelength, self.beam, self.atomic_data)
+
+        # spectrum parameters
+        min_wavelength = wavelength - 3
+        max_wavelength = wavelength + 3
+        bins = 512
+        point = Point3D(0.5, 0.5, 0.5)
+        direction = Vector3D(-1, 1, 0) / np.sqrt(2)
+        beam_direction = self.beam.direction(point.x, point.y, point.z)
+        beam_velocity = beam_direction.normalise() * np.sqrt(2 * self.beam.energy * ELEMENTARY_CHARGE / ATOMIC_MASS)
+
+        # obtaining spectrum
+        radiance = 1.0
+        spectrum = {}
+        for pol in ('no', 'pi', 'sigma'):
+            spectrum[pol] = Spectrum(min_wavelength, max_wavelength, bins)
+            mse_line.polarisation = pol
+            spectrum[pol] = mse_line.add_line(radiance, point, point, beam_velocity, direction, spectrum[pol])
+
+        # validating
+
+        # calculate Stark splitting
+        b_field = self.plasma.b_field(point.x, point.y, point.z)
+        e_field = beam_velocity.cross(b_field)
+        e_field_magn = e_field.length
+        STARK_ENERGY_SPLITTING_FACTOR = 3 * VACUUM_PERMITTIVITY * PLANCK_CONSTANT**2 / (2 * np.pi * ELECTRON_REST_MASS * ELEMENTARY_CHARGE**2)
+        stark_split = STARK_ENERGY_SPLITTING_FACTOR * e_field_magn
+
+        # obtain Stark relative ratios
+        ne = self.plasma.electron_distribution.density(point.x, point.y, point.z)
+        stark_structure = self.atomic_data.stark_structure(line)
+        ratios = stark_structure(self.beam.energy, ne, b_field.length)
+        indx_sigma, = np.where(stark_structure.polarisation == 1)
+        weight = np.ones(indx_sigma.size)
+        weight[stark_structure.index[indx_sigma] > 0] = 2.
+        sigma_to_total = (ratios[indx_sigma] * weight).sum()
+
+        # calculate emission line central wavelength, doppler shifted along observation direction
+        shifted_wavelength = wavelength * (1 + beam_velocity.dot(direction.normalise()) / SPEED_OF_LIGHT)
+        photon_energy = HC_EV_NM / shifted_wavelength
+
+        # calculate doppler broadening
+        beam_ion_mass = self.beam.element.atomic_weight
+        beam_temperature = self.beam.temperature
+        sigma = np.sqrt(beam_temperature * ELEMENTARY_CHARGE / (beam_ion_mass * ATOMIC_MASS)) * wavelength / SPEED_OF_LIGHT
+        temp = 1. / (np.sqrt(2.) * sigma)
+
+        # coefficients for intensities parallel and perpendicular to electric field
+        cos_sqr = e_field.normalise().dot(direction.normalise())**2
+        sin_sqr = 1. - cos_sqr
+
+        # calculate relative intensities of sigma and pi lines
+        intensity_sig = (sin_sqr + cos_sqr / sigma_to_total) * radiance
+        intensity_pi = sin_sqr * radiance
+
+        wavelengths, delta = np.linspace(min_wavelength, max_wavelength, bins + 1, retstep=True)
+        test_spectrum = {'sigma': 0, 'pi': 0}
+
+        for indx, pol, ratio in zip(stark_structure.index, stark_structure.polarisation, ratios):
+            if pol == 1:  # sigma
+                if indx == 0:
+                    erfs = erf((wavelengths - shifted_wavelength) * temp)
+                    test_spectrum['sigma'] += 0.5 * intensity_sig * ratio * (erfs[1:] - erfs[:-1]) / delta
+                else:
+                    erfs = erf((wavelengths - HC_EV_NM / (photon_energy + indx * stark_split)) * temp)
+                    test_spectrum['sigma'] += 0.5 * intensity_sig * ratio * (erfs[1:] - erfs[:-1]) / delta
+                    erfs = erf((wavelengths - HC_EV_NM / (photon_energy - indx * stark_split)) * temp)
+                    test_spectrum['sigma'] += 0.5 * intensity_sig * ratio * (erfs[1:] - erfs[:-1]) / delta
+            else:  # pi
+                erfs = erf((wavelengths - HC_EV_NM / (photon_energy + indx * stark_split)) * temp)
+                test_spectrum['pi'] += 0.5 * intensity_pi * ratio * (erfs[1:] - erfs[:-1]) / delta
+                erfs = erf((wavelengths - HC_EV_NM / (photon_energy - indx * stark_split)) * temp)
+                test_spectrum['pi'] += 0.5 * intensity_pi * ratio * (erfs[1:] - erfs[:-1]) / delta
+
+        test_spectrum['no'] = test_spectrum['pi'] + test_spectrum['sigma']
+
+        for pol in ('no', 'pi', 'sigma'):
+            for i in range(bins):
+                self.assertAlmostEqual(test_spectrum[pol][i], spectrum[pol].samples[i], delta=1e-10,
+                                       msg='BeamEmissionMultiplet.add_line() method gives a wrong value at {} nm.'.format(wavelengths[i]))
 
 
 if __name__ == '__main__':